Moved PhotoTile to separate file
